tcpip_suite/__link_layer/pkg/security/auth8021x.go
// pkg/security/auth8021x.go

package security

import (
    "errors"
    "fmt"
    "os/exec"

    "link_layer/pkg/utils"
)

// Auth8021X represents the 802.1X authentication configuration.
type Auth8021X struct {
    Enabled       bool
    EAPMethod     string
    Identity      string
    Certificate   string
    PrivateKey    string
    CACertificate string
}

// Initialize8021X initializes 802.1X authentication on a given interface.
func (auth *Auth8021X) Initialize(interfaceName string) error {
    if !auth.Enabled {
        utils.Logger.Println("802.1X Authentication is disabled.")
        return nil
    }

    // Validate EAP Method.
    if auth.EAPMethod != "PEAP" && auth.EAPMethod != "EAP-TLS" {
        return errors.New("unsupported EAP method")
    }

    // Configure 802.1X using system commands or a supplicant like wpa_supplicant.
    // Example using wpa_supplicant:
    /*
        wpaConf := fmt.Sprintf(`
        ctrl_interface=DIR=/var/run/wpa_supplicant GROUP=netdev
        network={
            key_mgmt=IEEE8021X
            eap=%s
            identity="%s"
            %s
            ca_cert="%s"
            client_cert="%s"
            private_key="%s"
        }`, auth.EAPMethod, auth.Identity, 
            func() string {
                if auth.EAPMethod == "EAP-TLS" {
                    return fmt.Sprintf(`client_cert="%s"
            private_key="%s"`, auth.Certificate, auth.PrivateKey)
                }
                return ""
            }(),
            auth.CACertificate, auth.Certificate, auth.PrivateKey)

        err := ioutil.WriteFile("/etc/wpa_supplicant/wpa_supplicant.conf", []byte(wpaConf), 0644)
        if err != nil {
            return err
        }

        cmd := exec.Command("wpa_supplicant", "-B", "-i", interfaceName, "-c", "/etc/wpa_supplicant/wpa_supplicant.conf")
        output, err := cmd.CombinedOutput()
        if err != nil {
            return fmt.Errorf("failed to start wpa_supplicant: %v, output: %s", err, string(output))
        }
    */

    // Placeholder: Log intended action.
    utils.Logger.Printf("Initialized 802.1X Authentication on interface %s with EAP method %s", interfaceName, auth.EAPMethod)
    return nil
}

// Remove8021X removes 802.1X authentication from the interface.
func (auth *Auth8021X) Remove8021X(interfaceName string) error {
    if !auth.Enabled {
        utils.Logger.Println("802.1X Authentication is disabled.")
        return nil
    }

    // Example: Stop wpa_supplicant.
    /*
        cmd := exec.Command("killall", "wpa_supplicant")
        output, err := cmd.CombinedOutput()
        if err != nil {
            return fmt.Errorf("failed to stop wpa_supplicant: %v, output: %s", err, string(output))
        }
    */

    // Placeholder: Log intended action.
    utils.Logger.Printf("Removed 802.1X Authentication from interface %s", interfaceName)
    return nil
}